set up matches fragment
