+++ app/src/main/res/layout/match.xml
@@ -4,10 +4,10 @@
     android:layout_width="match_parent" android:layout_height="match_parent"
     android:orientation="horizontal">
 
-    <ImageView
+    <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/imageView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_width="75dp"
+        android:layout_height="110dp"
         android:layout_weight=".50"
         app:layout_constraintLeft_toLeftOf="parent"
         app:srcCompat="@mipmap/gitinder_logo"/>
